requirements.txt
discord.py==2.2.2
openai==0.27.4
Pillow==9.5.0
Pillow==9.5.0
pyautogui==0.9.53
pyperclip==1.8.2
python-dotenv==1.0.0
Requests==2.28.2